Added `QrCodeSensorUseCaseTest`
