+++ core/test/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
     api(libs.androidx.arch.core.testing)
     api(libs.mockk.jvm)
     api(libs.kotlin.test)
+    api(libs.kotlin.reflect)
     api(libs.turbine)
 
     implementation(project(":core:database"))
