+++ feature/qrcode/build.gradle.kts
@@ -18,6 +18,8 @@ dependencies {
     implementation(project(":core:common"))
     implementation(project(":core:ui"))
 
+    testImplementation(project(":core:test"))
+
     implementation(libs.google.mlkit.barecode)
     implementation(libs.androidx.camera2.core)
     implementation(libs.androidx.camera2.lifecycle)
